+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 3.1.1rc: 2024-10-25
 
 * Add tinyMCE plugin that adds helper class to elements (T-17631)
+* Use eager loading instead of lazy if `native_lazyload_tag` is used in the first block (T-17629)
 
 ### 3.1.0: 2024-10-10